requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.15
Werkzeug==2.3.4
